test: improve test case
