+++ tests/unit/test_ho_to_raw_codemod.py
@@ -1,7 +1,5 @@
 import ast
-from pathlib import Path
 
-import pytest
 from freqtrade.configuration import Configuration
 from freqtrade.resolvers import StrategyResolver
 
@@ -123,18 +121,16 @@ def test_transform_code_syntax():
 def test_transform_and_replace_references():
     test_data_dir = REPO_ROOT.joinpath("tests/unit/test_data")
     config = Configuration.from_files([])
-    config["strategy"] = "Strategy1"
+    config["strategy"] = "Strategy"
     config["strategy_path"] = test_data_dir.joinpath("input")
     strategy = StrategyResolver.load_strategy(config)
     source_input = ""
-    with open(config["strategy_path"].joinpath("Strategy1.py")) as f:
+    with open(config["strategy_path"].joinpath("Strategy.py")) as f:
         source_input = f.read()
     transformed_input = transform_code(replace_all_references(strategy, source_input))
 
     expected_source = ""
-    with open(test_data_dir.joinpath("output/Strategy1.py")) as f:
+    with open(test_data_dir.joinpath("output/Strategy.py")) as f:
         expected_source = f.read()
 
     assert transformed_input.strip() == expected_source.strip()
-
-    # assert transformed_input.strip() == expected_output.strip()
